+++ workbench/source_gen/jetbrains/mps/ide/actions/Mac_KeymapChanges.java
@@ -46,7 +46,7 @@ public class Mac_KeymapChanges extends BaseKeymapChanges {
 
     public List<Shortcut> getShortcutsFor(Object... objects) {
       List<Shortcut> _result = new ArrayList<Shortcut>();
-      _result.add(new KeyboardShortcut(KeyStroke.getKeyStroke("ctrl " + objects[0]), null));
+      _result.add(new KeyboardShortcut(KeyStroke.getKeyStroke("ctrl " + ((Integer) objects[0])), null));
       return _result;
     }
   }
